--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPUTensorAlloc.cpp
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPUTensorAlloc.cpp
@@ -23,7 +23,9 @@ static bool contractOpFilter(Operation *op) {
   auto linalgOp = dyn_cast<linalg::LinalgOp>(op);
   if (!linalgOp) return false;
 
-  // Can't promote dynamic shapes.
+  // The workgroup specialization already makes static shapes available for the
+  // main tile part and makes the partial tile computation small, so promoting
+  // to shared memory for the partial tile actually hurts the performance.
   if (linalgOp.hasDynamicShape()) return false;
 
   SmallVector<unsigned> dims;