Specialization helped the performance too and now the sharemem promotion actually slows down the partial tile portion. Needs a broad evaluation.

For example, distillbert shows improvement from 6.21ms to 5.29ms.