@okkwon 2022-10-14
Got a perf result on A5000. Except for `layoutlm`, overall it improves the performance. 

| Benchmark | Before(ms) | After(ms) | Speedup |
|-----------|------------|-----------|---------|
|albert-base-v2 | 7.26 | 7.22 | 1.01 |
|bert-base-uncased | 17.30 | 16.20 | 1.07 |
|camembert-base | 17.70 | 16.70 | 1.06 |
|dbmdz/convert-base-turkish-cased | 14.60 | 14.70 | 0.99 |
|distilbert-base-uncased | 14.6 | 13.5 | 1.08 |
|gooble/electra-small-discriminator | 4 | 4 | 1 |
|gooble/mobilebert-uncased | 15.20 | 14.70 | 1.03 |
|microsoft/layoutlm-base-uncased | 14.60 | 16.9 | 0.86 |
|microsoft/MiniLM-L12-H384-uncased | 2.92 | 2.94 | 0.99 |
|resnet50 | 8.68 | 8.71 | 1 |
|roberta-base | 27.80 | 27.10 | 1.03|