@okkwon 2022-10-14
Created https://github.com/iree-org/iree/issues/10802 to further investigate the layoutlm regression.